Moved Promise.any into a separate module, added it to index.js & polyfill.js
