--- src/aggregateError.js
@@ -1,10 +1,0 @@
-/**
- * @constructor
- */
-function AggregateError(errors, message) {
-  (this.name = 'AggregateError'), (this.errors = errors);
-  this.message = message || '';
-}
-AggregateError.prototype = Error.prototype;
-
-export default AggregateError;
