+++ src/index.js
@@ -1,6 +1,6 @@
 import promiseFinally from './finally';
 import allSettled from './allSettled';
-import AggregateError from './aggregateError';
+import any from './any';
 
 // Store setTimeout reference so promise-polyfill will be unaffected by
 // other code modifying setTimeout (like sinon.useFakeTimers())
@@ -205,37 +205,7 @@ Promise.all = function(arr) {
   });
 };
 
-Promise.any = function(arr) {
-  return new Promise(function(resolve, reject) {
-    if (!isArray(arr)) {
-      return reject(new TypeError('Promise.any accepts an array'));
-    }
-
-    var args = Array.prototype.slice.call(arr);
-    if (args.length === 0) return reject();
-
-    var rejectionReasons = [];
-    for (var i = 0; i < args.length; i++) {
-      try {
-        Promise.resolve(args[i])
-          .then(resolve)
-          .catch(function(error) {
-            rejectionReasons.push(error);
-            if (rejectionReasons.length === args.length) {
-              reject(
-                new AggregateError(
-                  rejectionReasons,
-                  'All promises were rejected'
-                )
-              );
-            }
-          });
-      } catch (ex) {
-        reject(ex);
-      }
-    }
-  });
-};
+Promise.any = any;
 
 Promise.allSettled = allSettled;
